requirements.txt
aiofiles==23.2.1
boto3==1.38.15
botocore==1.38.15
certifi==2025.4.26
charset-normalizer==3.4.2
crcmod==1.7
dotenv==0.9.9
idna==3.10
jmespath==1.0.1
ks3sdk==1.12.1
pycryptodome==3.20.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
s3transfer==0.12.0
six==1.17.0
urllib3==2.4.0
